Add Spring Boot Database and Java Form applications

+++ src/main/java/com/oguz/springnews/repository/NewsRepository.java
@@ -1,8 +1,4 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+
 package com.oguz.springnews.repository;
 
 import com.oguz.springnews.entity.News;
@@ -10,10 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 /**
- *
- * @author oguz
+ * Yazilim Laboratuvari II Proje 2
+ * @author Oguz Aktas & Mert Var
  */
 @Repository
-public interface NewsRepository extends JpaRepository<News, Integer> {
+public interface NewsRepository extends JpaRepository<News, Long> {
     
 }
